test(cache): extract inline fixtures to dedicated files
AbstractPhpFileCacheBatchTest and BatchWriteBench each defined a second
AbstractPhpFileCache subclass inline. Move TestPhpFileCache and
BatchBenchCache to their own files to match the one-class-per-file
convention used elsewhere in the repo.
