+++ tests/Benchmark/FileCache/BatchBenchCache.php
@@ -1,0 +1,17 @@
+<?php
+
+declare(strict_types=1);
+
+namespace GacelaTest\Benchmark\FileCache;
+
+use Gacela\Framework\ClassResolver\Cache\AbstractPhpFileCache;
+
+final class BatchBenchCache extends AbstractPhpFileCache
+{
+    public const FILENAME = 'gacela-batch-bench.php';
+
+    protected function getCacheFilename(): string
+    {
+        return self::FILENAME;
+    }
+}
